requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
phonenumbers==8.13.51
pika==1.3.2
pydantic==2.10.2
pydantic-extra-types==2.10.1
pydantic_core==2.27.1
python-dotenv==1.0.1
python-telegram-bot==21.6
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
